CSP-267 conformed lat and long to only use nullable decimal

+++ src/SFA.DAS.Roatp.Domain/Interfaces/IProviderDetailsReadRepository.cs
@@ -6,11 +6,11 @@ namespace SFA.DAS.Roatp.Domain.Interfaces
 {
     public interface IProviderDetailsReadRepository
     {
-        Task<ProviderCourseDetailsModel> GetProviderDetailsWithDistance(int ukprn, int larsCode, double? lat, double? lon);
-        Task<List<ProviderCourseLocationDetailsModel>> GetProviderlocationDetailsWithDistance(int ukprn, int larsCode, double? lat, double? lon);
+        Task<ProviderCourseDetailsModel> GetProviderDetailsWithDistance(int ukprn, int larsCode, decimal? lat, decimal? lon);
+        Task<List<ProviderCourseLocationDetailsModel>> GetProviderlocationDetailsWithDistance(int ukprn, int larsCode, decimal? lat, decimal? lon);
 
-        Task<List<ProviderCourseDetailsSummaryModel>> GetAllProviderDetailsWithDistance(int larsCode, double? lat, double? lon);
-        Task<List<ProviderCourseLocationDetailsModel>> GetAllProviderlocationDetailsWithDistance(int larsCode, double? lat, double? lon);
+        Task<List<ProviderCourseDetailsSummaryModel>> GetAllProviderDetailsWithDistance(int larsCode, decimal? lat, decimal? lon);
+        Task<List<ProviderCourseLocationDetailsModel>> GetAllProviderlocationDetailsWithDistance(int larsCode, decimal? lat, decimal? lon);
 
 
     }
